working on excluding files
