+++ .circleci/config.yml
@@ -476,9 +476,7 @@ workflows:
                 name: Check code coverage 📈
                 # we will check the final coverage report
                 # to make sure it only has files we are interested in
-                command: |
-                  ../../node_modules/.bin/check-coverage main.js
-                  ../../node_modules/.bin/only-covered --from coverage/coverage-final.json main.js
+                command: npm run check
                 working_directory: examples/exclude-files
 
       - cypress/run:
