+++ examples/exclude-files/cypress.config.js
@@ -4,8 +4,10 @@ module.exports = defineConfig({
   e2e: {
     baseUrl: 'http://localhost:1234',
     specPattern: 'cypress/integration/*.js',
-    coverage: {
-      exclude: true,
+    env: {
+      coverage: {
+        exclude: true,
+      },
     },
     // We've imported your old cypress plugins here.
     // You may want to clean this up later by importing these.
